pyproject.toml
## pyproject.toml

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentlab"
version = "0.1.0"
description = "CLI framework for declarative AI agent blueprints"
readme = "README.md"
authors = [{name = "yzk"}]
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12.3",
  "rich>=13.7.1",
  "pydantic>=2.7.0",
  "PyYAML>=6.0.1",
  "httpx>=0.27.0",
  "orjson>=3.10.7",
]

[project.scripts]
agentlab = "agentlab.cli:app"
